+++ random.h
@@ -2,9 +2,9 @@
 #ifndef RANDOM_H_
 #define RANDOM_H_
 
-#include <cstddef>
-
 #include <random>
+
+#include <cstddef>
 
 #include "./global.h"
 
@@ -20,12 +20,12 @@ public:
 	 * Write into |output| |choices| random values based on |seed|.
 	 * Produces random floating-point values in the interval [0, 1)
 	 */
-	void GenerateChoices(int choices, float* output);
+	void GenerateChoices(int choices, double* output);
 
 	/*
 	 * Generate a floating-point value in the interval [0, 1).
 	 */
-	float next();
+	double next();
 
 	std::mt19937& engine();
 
@@ -33,7 +33,7 @@ public:
 	uint64_t seed() const;
 
 private:
-	std::uniform_real_distribution<float> distribution_;
+	std::uniform_real_distribution<double> distribution_;
 	std::mt19937 mersenne_engine_;  // Mersenne twister MT19937
 
 	uint64_t seed_;
@@ -42,6 +42,5 @@ private:
 };
 
 }  // namespace acs
-
 
 #endif  // RANDOM_H_
